LayoutTests/fast/canvas/webgl/constants-on-interface-expected.txt

WebGLRenderingContext.ACTIVE_ATTRIBUTES == 35721
WebGLRenderingContext.ACTIVE_TEXTURE == 34016
WebGLRenderingContext.ACTIVE_UNIFORMS == 35718
WebGLRenderingContext.ALIASED_LINE_WIDTH_RANGE == 33902
WebGLRenderingContext.ALIASED_POINT_SIZE_RANGE == 33901
WebGLRenderingContext.ALPHA == 6406
WebGLRenderingContext.ALPHA_BITS == 3413
WebGLRenderingContext.ALWAYS == 519
WebGLRenderingContext.ARRAY_BUFFER == 34962
WebGLRenderingContext.ARRAY_BUFFER_BINDING == 34964
WebGLRenderingContext.ATTACHED_SHADERS == 35717
WebGLRenderingContext.BACK == 1029
WebGLRenderingContext.BLEND == 3042
WebGLRenderingContext.BLEND_COLOR == 32773
WebGLRenderingContext.BLEND_DST_ALPHA == 32970
WebGLRenderingContext.BLEND_DST_RGB == 32968
WebGLRenderingContext.BLEND_EQUATION == 32777
WebGLRenderingContext.BLEND_EQUATION_ALPHA == 34877
WebGLRenderingContext.BLEND_EQUATION_RGB == 32777
WebGLRenderingContext.BLEND_SRC_ALPHA == 32971
WebGLRenderingContext.BLEND_SRC_RGB == 32969
WebGLRenderingContext.BLUE_BITS == 3412
WebGLRenderingContext.BOOL == 35670
WebGLRenderingContext.BOOL_VEC2 == 35671
WebGLRenderingContext.BOOL_VEC3 == 35672
WebGLRenderingContext.BOOL_VEC4 == 35673
WebGLRenderingContext.BROWSER_DEFAULT_WEBGL == 37444
WebGLRenderingContext.BUFFER_SIZE == 34660
WebGLRenderingContext.BUFFER_USAGE == 34661
WebGLRenderingContext.BYTE == 5120
WebGLRenderingContext.CCW == 2305
WebGLRenderingContext.CLAMP_TO_EDGE == 33071
WebGLRenderingContext.COLOR_ATTACHMENT0 == 36064
WebGLRenderingContext.COLOR_BUFFER_BIT == 16384
WebGLRenderingContext.COLOR_CLEAR_VALUE == 3106
WebGLRenderingContext.COLOR_WRITEMASK == 3107
WebGLRenderingContext.COMPILE_STATUS == 35713
WebGLRenderingContext.COMPRESSED_TEXTURE_FORMATS == 34467
WebGLRenderingContext.CONSTANT_ALPHA == 32771
WebGLRenderingContext.CONSTANT_COLOR == 32769
WebGLRenderingContext.CONTEXT_LOST_WEBGL == 37442
WebGLRenderingContext.CULL_FACE == 2884
WebGLRenderingContext.CULL_FACE_MODE == 2885
WebGLRenderingContext.CURRENT_PROGRAM == 35725
WebGLRenderingContext.CURRENT_VERTEX_ATTRIB == 34342
WebGLRenderingContext.CW == 2304
WebGLRenderingContext.DECR == 7683
WebGLRenderingContext.DECR_WRAP == 34056
WebGLRenderingContext.DELETE_STATUS == 35712
WebGLRenderingContext.DEPTH_ATTACHMENT == 36096
WebGLRenderingContext.DEPTH_BITS == 3414
WebGLRenderingContext.DEPTH_BUFFER_BIT == 256
WebGLRenderingContext.DEPTH_CLEAR_VALUE == 2931
WebGLRenderingContext.DEPTH_COMPONENT == 6402
WebGLRenderingContext.DEPTH_COMPONENT16 == 33189
WebGLRenderingContext.DEPTH_FUNC == 2932
WebGLRenderingContext.DEPTH_RANGE == 2928
WebGLRenderingContext.DEPTH_STENCIL == 34041
WebGLRenderingContext.DEPTH_STENCIL_ATTACHMENT == 33306
WebGLRenderingContext.DEPTH_TEST == 2929
WebGLRenderingContext.DEPTH_WRITEMASK == 2930
WebGLRenderingContext.DITHER == 3024
WebGLRenderingContext.DONT_CARE == 4352
WebGLRenderingContext.DST_ALPHA == 772
WebGLRenderingContext.DST_COLOR == 774
WebGLRenderingContext.DYNAMIC_DRAW == 35048
WebGLRenderingContext.ELEMENT_ARRAY_BUFFER == 34963
WebGLRenderingContext.ELEMENT_ARRAY_BUFFER_BINDING == 34965
WebGLRenderingContext.EQUAL == 514
WebGLRenderingContext.FASTEST == 4353
WebGLRenderingContext.FLOAT == 5126
WebGLRenderingContext.FLOAT_MAT2 == 35674
WebGLRenderingContext.FLOAT_MAT3 == 35675
WebGLRenderingContext.FLOAT_MAT4 == 35676
WebGLRenderingContext.FLOAT_VEC2 == 35664
WebGLRenderingContext.FLOAT_VEC3 == 35665
WebGLRenderingContext.FLOAT_VEC4 == 35666
WebGLRenderingContext.FRAGMENT_SHADER == 35632
WebGLRenderingContext.FRAMEBUFFER == 36160
WebGLRenderingContext.FRAMEBUFFER_ATTACHMENT_OBJECT_NAME == 36049
WebGLRenderingContext.FRAMEBUFFER_ATTACHMENT_OBJECT_TYPE == 36048
WebGLRenderingContext.FRAMEBUFFER_ATTACHMENT_TEXTURE_CUBE_MAP_FACE == 36051
WebGLRenderingContext.FRAMEBUFFER_ATTACHMENT_TEXTURE_LEVEL == 36050
WebGLRenderingContext.FRAMEBUFFER_BINDING == 36006
WebGLRenderingContext.FRAMEBUFFER_COMPLETE == 36053
WebGLRenderingContext.FRAMEBUFFER_INCOMPLETE_ATTACHMENT == 36054
WebGLRenderingContext.FRAMEBUFFER_INCOMPLETE_DIMENSIONS == 36057
WebGLRenderingContext.FRAMEBUFFER_INCOMPLETE_MISSING_ATTACHMENT == 36055
WebGLRenderingContext.FRAMEBUFFER_UNSUPPORTED == 36061
WebGLRenderingContext.FRONT == 1028
WebGLRenderingContext.FRONT_AND_BACK == 1032
WebGLRenderingContext.FRONT_FACE == 2886
WebGLRenderingContext.FUNC_ADD == 32774
WebGLRenderingContext.FUNC_REVERSE_SUBTRACT == 32779
WebGLRenderingContext.FUNC_SUBTRACT == 32778
WebGLRenderingContext.GENERATE_MIPMAP_HINT == 33170
WebGLRenderingContext.GEQUAL == 518
WebGLRenderingContext.GREATER == 516
WebGLRenderingContext.GREEN_BITS == 3411
WebGLRenderingContext.HIGH_FLOAT == 36338
WebGLRenderingContext.HIGH_INT == 36341
WebGLRenderingContext.IMPLEMENTATION_COLOR_READ_FORMAT == 35739
WebGLRenderingContext.IMPLEMENTATION_COLOR_READ_TYPE == 35738
WebGLRenderingContext.INCR == 7682
WebGLRenderingContext.INCR_WRAP == 34055
WebGLRenderingContext.INT == 5124
WebGLRenderingContext.INT_VEC2 == 35667
WebGLRenderingContext.INT_VEC3 == 35668
WebGLRenderingContext.INT_VEC4 == 35669
WebGLRenderingContext.INVALID_ENUM == 1280
WebGLRenderingContext.INVALID_FRAMEBUFFER_OPERATION == 1286
WebGLRenderingContext.INVALID_OPERATION == 1282
WebGLRenderingContext.INVALID_VALUE == 1281
WebGLRenderingContext.INVERT == 5386
WebGLRenderingContext.KEEP == 7680
WebGLRenderingContext.LEQUAL == 515
WebGLRenderingContext.LESS == 513
WebGLRenderingContext.LINEAR == 9729
WebGLRenderingContext.LINEAR_MIPMAP_LINEAR == 9987
WebGLRenderingContext.LINEAR_MIPMAP_NEAREST == 9985
WebGLRenderingContext.LINES == 1
WebGLRenderingContext.LINE_LOOP == 2
WebGLRenderingContext.LINE_STRIP == 3
WebGLRenderingContext.LINE_WIDTH == 2849
WebGLRenderingContext.LINK_STATUS == 35714
WebGLRenderingContext.LOW_FLOAT == 36336
WebGLRenderingContext.LOW_INT == 36339
WebGLRenderingContext.LUMINANCE == 6409
WebGLRenderingContext.LUMINANCE_ALPHA == 6410
WebGLRenderingContext.MAX_COMBINED_TEXTURE_IMAGE_UNITS == 35661
WebGLRenderingContext.MAX_CUBE_MAP_TEXTURE_SIZE == 34076
WebGLRenderingContext.MAX_FRAGMENT_UNIFORM_VECTORS == 36349
WebGLRenderingContext.MAX_RENDERBUFFER_SIZE == 34024
WebGLRenderingContext.MAX_TEXTURE_IMAGE_UNITS == 34930
WebGLRenderingContext.MAX_TEXTURE_SIZE == 3379
WebGLRenderingContext.MAX_VARYING_VECTORS == 36348
WebGLRenderingContext.MAX_VERTEX_ATTRIBS == 34921
WebGLRenderingContext.MAX_VERTEX_TEXTURE_IMAGE_UNITS == 35660
WebGLRenderingContext.MAX_VERTEX_UNIFORM_VECTORS == 36347
WebGLRenderingContext.MAX_VIEWPORT_DIMS == 3386
WebGLRenderingContext.MEDIUM_FLOAT == 36337
WebGLRenderingContext.MEDIUM_INT == 36340
WebGLRenderingContext.MIRRORED_REPEAT == 33648
WebGLRenderingContext.NEAREST == 9728
WebGLRenderingContext.NEAREST_MIPMAP_LINEAR == 9986
WebGLRenderingContext.NEAREST_MIPMAP_NEAREST == 9984
WebGLRenderingContext.NEVER == 512
WebGLRenderingContext.NICEST == 4354
WebGLRenderingContext.NONE == 0
WebGLRenderingContext.NOTEQUAL == 517
WebGLRenderingContext.NO_ERROR == 0
WebGLRenderingContext.ONE == 1
WebGLRenderingContext.ONE_MINUS_CONSTANT_ALPHA == 32772
WebGLRenderingContext.ONE_MINUS_CONSTANT_COLOR == 32770
WebGLRenderingContext.ONE_MINUS_DST_ALPHA == 773
WebGLRenderingContext.ONE_MINUS_DST_COLOR == 775
WebGLRenderingContext.ONE_MINUS_SRC_ALPHA == 771
WebGLRenderingContext.ONE_MINUS_SRC_COLOR == 769
WebGLRenderingContext.OUT_OF_MEMORY == 1285
WebGLRenderingContext.PACK_ALIGNMENT == 3333
WebGLRenderingContext.POINTS == 0
WebGLRenderingContext.POLYGON_OFFSET_FACTOR == 32824
WebGLRenderingContext.POLYGON_OFFSET_FILL == 32823
WebGLRenderingContext.POLYGON_OFFSET_UNITS == 10752
WebGLRenderingContext.RED_BITS == 3410
WebGLRenderingContext.RENDERBUFFER == 36161
WebGLRenderingContext.RENDERBUFFER_ALPHA_SIZE == 36179
WebGLRenderingContext.RENDERBUFFER_BINDING == 36007
WebGLRenderingContext.RENDERBUFFER_BLUE_SIZE == 36178
WebGLRenderingContext.RENDERBUFFER_DEPTH_SIZE == 36180
WebGLRenderingContext.RENDERBUFFER_GREEN_SIZE == 36177
WebGLRenderingContext.RENDERBUFFER_HEIGHT == 36163
WebGLRenderingContext.RENDERBUFFER_INTERNAL_FORMAT == 36164
WebGLRenderingContext.RENDERBUFFER_RED_SIZE == 36176
WebGLRenderingContext.RENDERBUFFER_STENCIL_SIZE == 36181
WebGLRenderingContext.RENDERBUFFER_WIDTH == 36162
WebGLRenderingContext.RENDERER == 7937
WebGLRenderingContext.REPEAT == 10497
WebGLRenderingContext.REPLACE == 7681
WebGLRenderingContext.RGB == 6407
WebGLRenderingContext.RGB565 == 36194
WebGLRenderingContext.RGB5_A1 == 32855
WebGLRenderingContext.RGBA == 6408
WebGLRenderingContext.RGBA4 == 32854
WebGLRenderingContext.SAMPLER_2D == 35678
WebGLRenderingContext.SAMPLER_CUBE == 35680
WebGLRenderingContext.SAMPLES == 32937
WebGLRenderingContext.SAMPLE_ALPHA_TO_COVERAGE == 32926
WebGLRenderingContext.SAMPLE_BUFFERS == 32936
WebGLRenderingContext.SAMPLE_COVERAGE == 32928
WebGLRenderingContext.SAMPLE_COVERAGE_INVERT == 32939
WebGLRenderingContext.SAMPLE_COVERAGE_VALUE == 32938
WebGLRenderingContext.SCISSOR_BOX == 3088
WebGLRenderingContext.SCISSOR_TEST == 3089
WebGLRenderingContext.SHADER_TYPE == 35663
WebGLRenderingContext.SHADING_LANGUAGE_VERSION == 35724
WebGLRenderingContext.SHORT == 5122
WebGLRenderingContext.SRC_ALPHA == 770
WebGLRenderingContext.SRC_ALPHA_SATURATE == 776
WebGLRenderingContext.SRC_COLOR == 768
WebGLRenderingContext.STATIC_DRAW == 35044
WebGLRenderingContext.STENCIL_ATTACHMENT == 36128
WebGLRenderingContext.STENCIL_BACK_FAIL == 34817
WebGLRenderingContext.STENCIL_BACK_FUNC == 34816
WebGLRenderingContext.STENCIL_BACK_PASS_DEPTH_FAIL == 34818
WebGLRenderingContext.STENCIL_BACK_PASS_DEPTH_PASS == 34819
WebGLRenderingContext.STENCIL_BACK_REF == 36003
WebGLRenderingContext.STENCIL_BACK_VALUE_MASK == 36004
WebGLRenderingContext.STENCIL_BACK_WRITEMASK == 36005
WebGLRenderingContext.STENCIL_BITS == 3415
WebGLRenderingContext.STENCIL_BUFFER_BIT == 1024
WebGLRenderingContext.STENCIL_CLEAR_VALUE == 2961
WebGLRenderingContext.STENCIL_FAIL == 2964
WebGLRenderingContext.STENCIL_FUNC == 2962
WebGLRenderingContext.STENCIL_INDEX == 6401
WebGLRenderingContext.STENCIL_INDEX8 == 36168
WebGLRenderingContext.STENCIL_PASS_DEPTH_FAIL == 2965
WebGLRenderingContext.STENCIL_PASS_DEPTH_PASS == 2966
WebGLRenderingContext.STENCIL_REF == 2967
WebGLRenderingContext.STENCIL_TEST == 2960
WebGLRenderingContext.STENCIL_VALUE_MASK == 2963
WebGLRenderingContext.STENCIL_WRITEMASK == 2968
WebGLRenderingContext.STREAM_DRAW == 35040
WebGLRenderingContext.SUBPIXEL_BITS == 3408
WebGLRenderingContext.TEXTURE == 5890
WebGLRenderingContext.TEXTURE0 == 33984
WebGLRenderingContext.TEXTURE1 == 33985
WebGLRenderingContext.TEXTURE10 == 33994
WebGLRenderingContext.TEXTURE11 == 33995
WebGLRenderingContext.TEXTURE12 == 33996
WebGLRenderingContext.TEXTURE13 == 33997
WebGLRenderingContext.TEXTURE14 == 33998
WebGLRenderingContext.TEXTURE15 == 33999
WebGLRenderingContext.TEXTURE16 == 34000
WebGLRenderingContext.TEXTURE17 == 34001
WebGLRenderingContext.TEXTURE18 == 34002
WebGLRenderingContext.TEXTURE19 == 34003
WebGLRenderingContext.TEXTURE2 == 33986
WebGLRenderingContext.TEXTURE20 == 34004
WebGLRenderingContext.TEXTURE21 == 34005
WebGLRenderingContext.TEXTURE22 == 34006
WebGLRenderingContext.TEXTURE23 == 34007
WebGLRenderingContext.TEXTURE24 == 34008
WebGLRenderingContext.TEXTURE25 == 34009
WebGLRenderingContext.TEXTURE26 == 34010
WebGLRenderingContext.TEXTURE27 == 34011
WebGLRenderingContext.TEXTURE28 == 34012
WebGLRenderingContext.TEXTURE29 == 34013
WebGLRenderingContext.TEXTURE3 == 33987
WebGLRenderingContext.TEXTURE30 == 34014
WebGLRenderingContext.TEXTURE31 == 34015
WebGLRenderingContext.TEXTURE4 == 33988
WebGLRenderingContext.TEXTURE5 == 33989
WebGLRenderingContext.TEXTURE6 == 33990
WebGLRenderingContext.TEXTURE7 == 33991
WebGLRenderingContext.TEXTURE8 == 33992
WebGLRenderingContext.TEXTURE9 == 33993
WebGLRenderingContext.TEXTURE_2D == 3553
WebGLRenderingContext.TEXTURE_BINDING_2D == 32873
WebGLRenderingContext.TEXTURE_BINDING_CUBE_MAP == 34068
WebGLRenderingContext.TEXTURE_CUBE_MAP == 34067
WebGLRenderingContext.TEXTURE_CUBE_MAP_NEGATIVE_X == 34070
WebGLRenderingContext.TEXTURE_CUBE_MAP_NEGATIVE_Y == 34072
WebGLRenderingContext.TEXTURE_CUBE_MAP_NEGATIVE_Z == 34074
WebGLRenderingContext.TEXTURE_CUBE_MAP_POSITIVE_X == 34069
WebGLRenderingContext.TEXTURE_CUBE_MAP_POSITIVE_Y == 34071
WebGLRenderingContext.TEXTURE_CUBE_MAP_POSITIVE_Z == 34073
WebGLRenderingContext.TEXTURE_MAG_FILTER == 10240
WebGLRenderingContext.TEXTURE_MIN_FILTER == 10241
WebGLRenderingContext.TEXTURE_WRAP_S == 10242
WebGLRenderingContext.TEXTURE_WRAP_T == 10243
WebGLRenderingContext.TRIANGLES == 4
WebGLRenderingContext.TRIANGLE_FAN == 6
WebGLRenderingContext.TRIANGLE_STRIP == 5
WebGLRenderingContext.UNPACK_ALIGNMENT == 3317
WebGLRenderingContext.UNPACK_COLORSPACE_CONVERSION_WEBGL == 37443
WebGLRenderingContext.UNPACK_FLIP_Y_WEBGL == 37440
WebGLRenderingContext.UNPACK_PREMULTIPLY_ALPHA_WEBGL == 37441
WebGLRenderingContext.UNSIGNED_BYTE == 5121
WebGLRenderingContext.UNSIGNED_INT == 5125
WebGLRenderingContext.UNSIGNED_SHORT == 5123
WebGLRenderingContext.UNSIGNED_SHORT_4_4_4_4 == 32819
WebGLRenderingContext.UNSIGNED_SHORT_5_5_5_1 == 32820
WebGLRenderingContext.UNSIGNED_SHORT_5_6_5 == 33635
WebGLRenderingContext.VALIDATE_STATUS == 35715
WebGLRenderingContext.VENDOR == 7936
WebGLRenderingContext.VERSION == 7938
WebGLRenderingContext.VERTEX_ATTRIB_ARRAY_BUFFER_BINDING == 34975
WebGLRenderingContext.VERTEX_ATTRIB_ARRAY_ENABLED == 34338
WebGLRenderingContext.VERTEX_ATTRIB_ARRAY_NORMALIZED == 34922
WebGLRenderingContext.VERTEX_ATTRIB_ARRAY_POINTER == 34373
WebGLRenderingContext.VERTEX_ATTRIB_ARRAY_SIZE == 34339
WebGLRenderingContext.VERTEX_ATTRIB_ARRAY_STRIDE == 34340
WebGLRenderingContext.VERTEX_ATTRIB_ARRAY_TYPE == 34341
WebGLRenderingContext.VERTEX_SHADER == 35633
WebGLRenderingContext.VIEWPORT == 2978
WebGLRenderingContext.ZERO == 0
